pyproject.toml
[project]
name = "cpython-interp-perf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "capstone>=5.0.5",
    "pyelftools>=0.32",
    "pyperformance>=1.11.0",
]
